+++ src/NumericalAlgorithms/LinearOperators/PartialDerivatives.hpp
@@ -11,6 +11,7 @@
 #include <string>
 
 #include "DataStructures/DataBox/PrefixHelpers.hpp"
+#include "DataStructures/DataBox/Prefixes.hpp"
 #include "DataStructures/DataBox/Tag.hpp"
 #include "DataStructures/DataBox/TagName.hpp"
 #include "DataStructures/Variables.hpp"
@@ -35,57 +36,6 @@ template <class TagList>
 struct Variables;
 }  // namespace Tags
 /// \endcond
-
-namespace Tags {
-/*!
- * \ingroup DataBoxTagsGroup
- * \brief Prefix indicating spatial derivatives
- *
- * Prefix indicating the spatial derivatives of a Tensor.
- *
- * \tparam Tag The tag to wrap
- * \tparam Dim The volume dim as a type (e.g. `tmpl::size_t<Dim>`)
- * \tparam Frame The frame of the derivative index
- *
- * \see Tags::DerivCompute
- */
-template <typename Tag, typename Dim, typename Frame, typename = std::nullptr_t>
-struct deriv;
-
-template <typename Tag, typename Dim, typename Frame>
-struct deriv<Tag, Dim, Frame, Requires<tt::is_a_v<Tensor, typename Tag::type>>>
-    : db::PrefixTag, db::SimpleTag {
-  using type =
-      TensorMetafunctions::prepend_spatial_index<typename Tag::type, Dim::value,
-                                                 UpLo::Lo, Frame>;
-  using tag = Tag;
-};
-
-/*!
- * \ingroup DataBoxTagsGroup
- * \brief Prefix indicating spacetime derivatives
- *
- * Prefix indicating the spacetime derivatives of a Tensor or that a Variables
- * contains spatial derivatives of Tensors.
- *
- * \tparam Tag The tag to wrap
- * \tparam Dim The volume dim as a type (e.g. `tmpl::size_t<Dim>`)
- * \tparam Frame The frame of the derivative index
- */
-template <typename Tag, typename Dim, typename Frame, typename = std::nullptr_t>
-struct spacetime_deriv;
-
-template <typename Tag, typename Dim, typename Frame>
-struct spacetime_deriv<Tag, Dim, Frame,
-                       Requires<tt::is_a_v<Tensor, typename Tag::type>>>
-    : db::PrefixTag, db::SimpleTag {
-  using type =
-      TensorMetafunctions::prepend_spacetime_index<typename Tag::type,
-                                                   Dim::value, UpLo::Lo, Frame>;
-  using tag = Tag;
-};
-
-}  // namespace Tags
 
 namespace partial_derivatives_detail {
 void apply_matrix_in_first_dim(double* result, const double* input,
